Move Persona creator into Persona class and add test
