+++ packages/sanes-chrome-extension/src/logic/persona/persona.spec.ts
@@ -1,0 +1,10 @@
+import { Persona } from './persona';
+import { describeWithChains } from '../../utils/test/testExecutor';
+
+describeWithChains('Persona', () => {
+  it('can be created', async () => {
+    const persona = await Persona.create();
+    expect(persona).toBeTruthy();
+    persona.destroy();
+  });
+});
