+++ packages/sanes-chrome-extension/src/logic/persona/persona.ts
@@ -1,14 +1,42 @@
 import { Amount } from '@iov/bcp';
 import { MultiChainSigner } from '@iov/core';
 
-import { AccountsManager, AccountInfo } from '../user';
+import { AccountsManager, AccountInfo, createUserProfile } from '../user';
+import { chainConnector, getRuntimeConfiguration } from '../config';
 
 export class Persona {
+  /**
+   * Creates a new Persona instance.
+   *
+   * This function does everything that cannot be done in a constructor
+   * (because a constructor is synchonous): reading configs, connecting to the network,
+   * creating accounts.
+   */
+  public static async create(): Promise<Persona> {
+    const { chains } = await getRuntimeConfiguration();
+
+    const profile = await createUserProfile();
+    const signer = new MultiChainSigner(profile);
+
+    // connect chains
+    for (const chain of chains) {
+      const connector = chainConnector(chain.codecType, chain.bootstrapNodes);
+      signer.addChain(connector);
+    }
+
+    const manager = new AccountsManager(profile, chains);
+
+    // Setup accounts
+    await manager.generateAccount(0);
+
+    return new Persona(signer, manager);
+  }
+
   private readonly signer: MultiChainSigner;
   private readonly accountsManager: AccountsManager;
 
   /** The given signer and accountsManager must share the same UserProfile */
-  public constructor(
+  private constructor(
     signer: MultiChainSigner,
     accountsManager: AccountsManager
   ) {
@@ -42,5 +70,9 @@ export class Persona {
       }
     }
     return out;
+  }
+
+  public destroy(): void {
+    this.signer.shutdown();
   }
 }
