+++ packages/sanes-chrome-extension/src/utils/test/testExecutor.ts
@@ -1,2 +1,6 @@
 export const skipTests = (envVar: string | undefined): boolean => !envVar;
 export const mayTestChains = skipTests(process.env.CHAINS_ENABLED) ? xit : it;
+
+export const describeWithChains = skipTests(process.env.CHAINS_ENABLED)
+  ? xdescribe
+  : describe;
